requirements.txt
pyserial
pytelegraf
